Unfinished checkout functionality adding. Small config and employee management fixes.

+++ src/main/java/com/heroku/spacey/services/impl/EmployeeServiceImpl.java
@@ -28,9 +28,6 @@ public class EmployeeServiceImpl implements EmployeeService {
 
         Map<String, String> filters = new HashMap<>();
 
-        int pageNum = DEFAULT_PAGE_NUM;
-        int pageSize = DEFAULT_PAGE_SIZE;
-
         if (isValid(role)) {
             filters.put("role", role);
         }
@@ -39,13 +36,16 @@ public class EmployeeServiceImpl implements EmployeeService {
             filters.put("status", status);
         }
 
+        int pageNum = DEFAULT_PAGE_NUM;
+        int pageSize = DEFAULT_PAGE_SIZE;
+
         if (isValid(page)) {
             String[] pageProps = page.split("-");
             pageNum = Integer.parseInt(pageProps[0]);
             pageSize = Integer.parseInt(pageProps[1]);
         }
 
-        if (isValid(searchPrompt) && searchPrompt.trim().length() >= 2) {
+        if (isValid(searchPrompt)) {
             filters.put("search", searchPrompt);
         }
 
@@ -56,31 +56,27 @@ public class EmployeeServiceImpl implements EmployeeService {
         return employeeDao.getById(loginId);
     }
 
-    public void createEmployee(EmployeeDto employeeDto) throws SQLException {
-        if (isValid(employeeDto.getEmail()) && isValid(employeeDto.getUserRole())
-                && isValid(employeeDto.getFirstName()) && isValid(employeeDto.getLastName())
-                && isValid(employeeDto.getStatus())) {
-            employeeDao.insert(employeeDto);
-        } else {
-            throw new SQLException("Fill in all mandatory fields.");
-        }
+    public void createEmployee(EmployeeDto employeeDto) throws IllegalArgumentException, SQLException {
+        employeeDao.insert(employeeDto);
     }
 
-    public int updateEmployee(EmployeeDto employeeDto) throws SQLException {
-        if (isValid(employeeDto.getEmail()) && isValid(employeeDto.getUserRole())
-                && isValid(employeeDto.getFirstName()) && isValid(employeeDto.getLastName())
-                && isValid(employeeDto.getStatus())) {
-            return employeeDao.update(employeeDto);
-        } else {
-            throw new SQLException("Fill in all mandatory fields.");
+    public int updateEmployee(EmployeeDto employeeDto) throws IllegalArgumentException, SQLException {
+        if (employeeDao.update(employeeDto) == 0) {
+            throw new NotFoundException("Haven't found employee with such ID.");
         }
+
+        return employeeDao.update(employeeDto);
     }
 
     public int deleteEmployee(int loginId) throws SQLException {
+        if (employeeDao.delete(loginId) == 0) {
+            throw new NotFoundException("Haven't found employee with such ID.");
+        }
+
         return employeeDao.delete(loginId);
     }
 
     private boolean isValid(String value) {
-        return value != null && !value.isEmpty();
+        return value != null && !value.isBlank();
     }
 }
